+++ packages/terra-functional-testing/src/services/wdio-visual-regression-service/modules/makeAreaScreenshot.js
@@ -1,7 +1,7 @@
 import fsExtra from 'fs-extra';
 import path from 'path';
 import uuidv4 from 'uuid/v4';
-import logger from '@wdio/logger';
+import { Logger } from '@cerner/terra-cli';
 
 import ScreenshotStrategyManager from '../utils/ScreenshotStrategyManager';
 import getScreenDimensions from '../scripts/getScreenDimensions';
@@ -12,12 +12,12 @@ import { cropImage, mergeImages } from '../utils/image';
 import ScreenDimension from '../utils/ScreenDimension';
 import normalizeScreenshot from '../utils/normalizeScreenshot';
 
-const log = logger('wdio-visual-regression-service:makeAreaScreenshot');
+const logger = new Logger('[wdio-visual-regression-service:makeAreaScreenshot]');
 const tmpDir = path.resolve(__dirname, '..', '..', 'tmp');
 
 async function storeScreenshot(browser, screenDimensions, cropDimensions, base64Screenshot, filePath) {
   const normalizedBase64Screenshot = await normalizeScreenshot(browser, screenDimensions, base64Screenshot);
-  log.info(
+  logger.verbose(
     'crop screenshot with width: %s, height: %s, offsetX: %s, offsetY: %s',
     cropDimensions.getWidth(),
     cropDimensions.getHeight(),
@@ -31,12 +31,12 @@ async function storeScreenshot(browser, screenDimensions, cropDimensions, base64
 }
 
 export default async function makeAreaScreenshot(browser, startX, startY, endX, endY) {
-  log.info('requested a screenshot for the following area: %j', {
+  logger.verbose('requested a screenshot for the following area: %j', {
     startX, startY, endX, endY,
   });
 
   const screenDimensions = await browser.execute(getScreenDimensions);
-  log.info('detected screenDimensions %j', screenDimensions);
+  logger.verbose('detected screenDimensions %j', screenDimensions);
   const screenDimension = new ScreenDimension(screenDimensions, browser);
 
   const screenshotStrategy = ScreenshotStrategyManager.getStrategy(browser, screenDimension);
@@ -52,7 +52,7 @@ export default async function makeAreaScreenshot(browser, startX, startY, endX, 
     const cropImages = [];
     const screenshotPromises = [];
 
-    log.info('set page height to %s px', screenDimension.getDocumentHeight());
+    logger.verbose('set page height to %s px', screenDimension.getDocumentHeight());
     await browser.execute(pageHeight, `${screenDimension.getDocumentHeight()}px`);
 
     let loop = false;
@@ -60,13 +60,13 @@ export default async function makeAreaScreenshot(browser, startX, startY, endX, 
       const {
         x, y, indexX, indexY,
       } = screenshotStrategy.getScrollPosition();
-      log.info('scroll to coordinates x: %s, y: %s for index x: %s, y: %s', x, y, indexX, indexY);
+      logger.verbose('scroll to coordinates x: %s, y: %s for index x: %s, y: %s', x, y, indexX, indexY);
       /* eslint-disable no-await-in-loop */
 
       await browser.execute(virtualScroll, x, y, false);
       await browser.pause(100);
 
-      log.info('take screenshot');
+      logger.verbose('take screenshot');
       const base64Screenshot = await browser.takeScreenshot();
       const cropDimensions = screenshotStrategy.getCropDimensions();
       const filePath = path.join(dir, `${indexY}-${indexX}.png`);
@@ -87,17 +87,17 @@ export default async function makeAreaScreenshot(browser, startX, startY, endX, 
     const [mergedBase64Screenshot] = await Promise.all([
       Promise.resolve().then(async () => {
         await Promise.all(screenshotPromises);
-        log.info('merge images together');
+        logger.verbose('merge images together');
         const mergedBase64Images = await mergeImages(cropImages);
-        log.info('remove temp dir');
+        logger.verbose('remove temp dir');
         await fsExtra.remove(dir);
         return mergedBase64Images;
       }),
       Promise.resolve().then(async () => {
-        log.info('reset page height');
+        logger.verbose('reset page height');
         await browser.execute(pageHeight, '');
 
-        log.info('revert scroll to x: %s, y: %s', 0, 0);
+        logger.verbose('revert scroll to x: %s, y: %s', 0, 0);
         await browser.execute(virtualScroll, 0, 0, true);
       }),
     ]);